refactor(cli): structural rework
* split cli logic into subdirectory with new structure (with cli handlers in sub dir). This moves most of main into one cli handler and the binary will impor the cli in order to run the app instead. This makes the logic a lot more consistent
* creates a _get_commands that makes a declarative spec/source of truth for the commands that can be used by the parser logic and run_command
  without mixing concerns and/or duplicating a lot of code.
* defers prepare_runtime (ensuring paths) after parse so it won't run for --help, --version or if the parser errors
* merge sub parser help and description (what is shown on the main help vs the dedicated sub parser help view)
