+++ pyproject.toml
@@ -141,7 +141,7 @@ select = [
 
 [tool.ruff.lint.flake8-tidy-imports.banned-api]
 "gi.repository" = {msg = "Import GI modules from ulauncher.gi, not gi.repository directly."}
-"ulauncher.ui" = {msg = "Import ulauncher.ui only from main.py - it pulls in GTK and must not leak into extension or CLI runtime."}
+"ulauncher.ui" = {msg = "Import ulauncher.ui only from ulauncher/cli/commands/app.py - it pulls in GTK and must not leak into extension or CLI runtime."}
 "functools.lru_cache" = {msg = "Use ulauncher.utils.lru_cache instead to preserve parameter types."}
 "functools.cache" = {msg = "Use ulauncher.utils.lru_cache instead to preserve parameter types."}
 "functools.cached_property" = {msg = "cached_property erases property types. If you need to use this, create a typed wrapper similar to ulauncher.utils.lru_cache"}
