--- ulauncher/cli.py
@@ -1,126 +1,0 @@
-from __future__ import annotations
-
-import argparse
-import sys
-from functools import partial
-
-from ulauncher import version
-from ulauncher.modes.extensions.extension_cli_handlers import (
-    install_extension,
-    list_active_extensions,
-    preview_extension,
-    uninstall_extension,
-    upgrade_extensions,
-)
-from ulauncher.utils.lru_cache import lru_cache
-
-
-class CLIArguments(argparse.Namespace):
-    daemon: bool
-    dev: bool  # deprecated
-    verbose: bool
-    hide_window: bool  # deprecated
-    no_extensions: bool  # deprecated
-    no_window: bool  # deprecated unreleased alias for hide_window (made more sense semantically)
-    no_window_shadow: bool  # deprecated
-
-
-@lru_cache(maxsize=None)
-def get_args() -> CLIArguments:
-    """Get the command line arguments for the current runtime."""
-    return parse(sys.argv[1:])
-
-
-def parse(input_args: list[str]) -> CLIArguments:
-    """Parse CLI arguments"""
-    # Python's argparse is very similar to Gtk.Application.add_main_option_entries,
-    # but GTK adds in their own options we don't want like --help-gtk --help-gapplication --help-all
-    parser = argparse.ArgumentParser(
-        usage="%(prog)s OPTIONS or ARGUMENT",
-        description="Ulauncher is a GTK application launcher with support for extensions, shortcuts (scripts), calculator, file browser and custom themes.",  # noqa: E501
-        add_help=False,
-    )
-    parser.add_argument("-h", "--help", action="help", help="Show this help message and exit")
-    parser.add_argument("-v", "--verbose", action="store_true", help="Show debug log messages")
-    parser.add_argument(
-        "--version",
-        action="version",
-        help="Show version number and exit",
-        version=f"Ulauncher {version}",
-    )
-    parser.add_argument(
-        "--daemon",
-        action="store_true",
-        help="Run Ulauncher as a background process, without initially showing the window",
-    )
-    # deprecated
-    parser.add_argument("--dev", action="store_true", help=argparse.SUPPRESS)
-    parser.add_argument("--hide-window", action="store_true", help=argparse.SUPPRESS)
-    parser.add_argument("--no-extensions", action="store_true", help=argparse.SUPPRESS)
-    parser.add_argument("--no-window", action="store_true", help=argparse.SUPPRESS)
-    parser.add_argument("--no-window-shadow", action="store_true", help=argparse.SUPPRESS)
-
-    subparsers = parser.add_subparsers(
-        prog="ulauncher",
-        title="commands",
-        description="Available commands",
-        dest="command",
-        help="Command help",
-    )
-
-    # Extension commands at top level
-    list_parser = subparsers.add_parser(
-        "extensions",
-        aliases=["e"],
-        help="List installed extensions",
-        description="List all installed extensions with their status and information",
-    )
-    list_parser.set_defaults(handler=partial(list_active_extensions, list_parser))
-
-    install_parser = subparsers.add_parser(
-        "install",
-        aliases=["i"],
-        help="Install an extension from URL",
-        description="Install an extension from a Git URL or local path",
-    )
-    install_parser.add_argument("input", help="Git URL or path of the extension to install")
-    install_parser.set_defaults(handler=partial(install_extension, install_parser))
-
-    uninstall_parser = subparsers.add_parser(
-        "uninstall",
-        aliases=["rm"],
-        help="Uninstall an extension",
-        description="Remove an installed extension by ID or URL",
-    )
-    uninstall_parser.add_argument("input", help="Extension ID or URL to uninstall")
-    uninstall_parser.set_defaults(handler=partial(uninstall_extension, uninstall_parser))
-
-    upgrade_parser = subparsers.add_parser(
-        "upgrade",
-        aliases=["up"],
-        help="Upgrade extensions",
-        description="Upgrade one or all installed extensions to their latest versions",
-    )
-    upgrade_parser.add_argument(
-        "input",
-        nargs=argparse.OPTIONAL,
-        help="Optional extension ID or URL to upgrade (upgrades all if not specified)",
-    )
-    upgrade_parser.set_defaults(handler=partial(upgrade_extensions, upgrade_parser))
-
-    preview_parser = subparsers.add_parser(
-        "preview",
-        aliases=["pr"],
-        help="Preview extension",
-        description="Starts extension from a local path for development and debugging purposes",
-    )
-    preview_parser.add_argument(
-        "--with-debugger",
-        action="store_true",
-        default=False,
-        help="Start the extension with remote Python debugger enabled",
-    )
-    preview_parser.add_argument("path", help="Path to the extension source directory")
-    preview_parser.set_defaults(handler=partial(preview_extension, preview_parser))
-
-    return parser.parse_args(args=input_args, namespace=CLIArguments())
